quash warnings

+++ m3-libs/m3core/src/Csupport/libgcc/libgcc.c
@@ -74,10 +74,10 @@ int  __cmpdi2 ( INT64 x , INT64 y) { return x < y ? 0 : x == y ? 1 : 2; }
 int __ucmpdi2 (UINT64 x, UINT64 y) { return x < y ? 0 : x == y ? 1 : 2; }
 
 /* INT64 to float (double int to single float) */
-float __floatdisf (INT64 x) { return x; }
+float __floatdisf (INT64 x) { return (float)x; }
 
 /* INT64 to double (double int to double float) */
-double __floatdidf (INT64 x) { return x; }
+double __floatdidf (INT64 x) { return (double)x; }
 
 /*
 count leading zeros?
